backend/requirements.txt
supabase==2.9.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
aiohttp==3.9.1
mailscout